+++ internal/config/config.go
@@ -19,7 +19,7 @@ import (
 	"github.com/influxdata/telegraf/plugins/serializers"
 
 	"github.com/influxdata/config"
-	"github.com/naoina/toml/ast"
+	"github.com/influxdata/toml/ast"
 )
 
 // Config specifies the URL/user/password for the database that telegraf
@@ -68,7 +68,7 @@ type AgentConfig struct {
 	// same time, which can have a measurable effect on the system.
 	CollectionJitter internal.Duration
 
-	// Interval at which to flush data
+	// FlushInterval is the Interval at which to flush data
 	FlushInterval internal.Duration
 
 	// FlushJitter Jitters the flush interval by a random amount.
@@ -81,6 +81,11 @@ type AgentConfig struct {
 	// will cache. The buffer is cleared when a successful write occurs. When
 	// full, the oldest metrics will be overwritten.
 	MetricBufferLimit int
+
+	// FlushBufferWhenFull tells Telegraf to flush the metric buffer whenever
+	// it fills up, regardless of FlushInterval. Setting this option to true
+	// does _not_ deactivate FlushInterval.
+	FlushBufferWhenFull bool
 
 	// TODO(cam): Remove UTC and Precision parameters, they are no longer
 	// valid for the agent config. Leaving them here for now for backwards-
@@ -128,9 +133,7 @@ func (c *Config) ListTags() string {
 	return strings.Join(tags, " ")
 }
 
-var header = `###############################################################################
-#                           Telegraf Configuration                            #
-###############################################################################
+var header = `# Telegraf Configuration
 
 # Telegraf is entirely plugin driven. All metrics are gathered from the
 # declared inputs, and sent to the declared outputs.
@@ -148,57 +151,56 @@ var header = `##################################################################
 
 # Configuration for telegraf agent
 [agent]
-  ### Default data collection interval for all inputs
+  ## Default data collection interval for all inputs
   interval = "10s"
-  ### Rounds collection interval to 'interval'
-  ### ie, if interval="10s" then always collect on :00, :10, :20, etc.
+  ## Rounds collection interval to 'interval'
+  ## ie, if interval="10s" then always collect on :00, :10, :20, etc.
   round_interval = true
 
-  ### Telegraf will cache metric_buffer_limit metrics for each output, and will
-  ### flush this buffer on a successful write.
-  metric_buffer_limit = 10000
+  ## Telegraf will cache metric_buffer_limit metrics for each output, and will
+  ## flush this buffer on a successful write.
+  metric_buffer_limit = 1000
+  ## Flush the buffer whenever full, regardless of flush_interval.
+  flush_buffer_when_full = true
 
-  ### Collection jitter is used to jitter the collection by a random amount.
-  ### Each plugin will sleep for a random time within jitter before collecting.
-  ### This can be used to avoid many plugins querying things like sysfs at the
-  ### same time, which can have a measurable effect on the system.
+  ## Collection jitter is used to jitter the collection by a random amount.
+  ## Each plugin will sleep for a random time within jitter before collecting.
+  ## This can be used to avoid many plugins querying things like sysfs at the
+  ## same time, which can have a measurable effect on the system.
   collection_jitter = "0s"
 
-  ### Default flushing interval for all outputs. You shouldn't set this below
-  ### interval. Maximum flush_interval will be flush_interval + flush_jitter
+  ## Default flushing interval for all outputs. You shouldn't set this below
+  ## interval. Maximum flush_interval will be flush_interval + flush_jitter
   flush_interval = "10s"
-  ### Jitter the flush interval by a random amount. This is primarily to avoid
-  ### large write spikes for users running a large number of telegraf instances.
-  ### ie, a jitter of 5s and interval 10s means flushes will happen every 10-15s
+  ## Jitter the flush interval by a random amount. This is primarily to avoid
+  ## large write spikes for users running a large number of telegraf instances.
+  ## ie, a jitter of 5s and interval 10s means flushes will happen every 10-15s
   flush_jitter = "0s"
 
-  ### Run telegraf in debug mode
+  ## Run telegraf in debug mode
   debug = false
-  ### Run telegraf in quiet mode
+  ## Run telegraf in quiet mode
   quiet = false
-  ### Override default hostname, if empty use os.Hostname()
+  ## Override default hostname, if empty use os.Hostname()
   hostname = ""
 
 
-###############################################################################
-#                                  OUTPUTS                                    #
-###############################################################################
+#
+# OUTPUTS:
+#
 
 `
 
 var pluginHeader = `
-
-###############################################################################
-#                                  INPUTS                                     #
-###############################################################################
-
+#
+# INPUTS:
+#
 `
 
 var serviceInputHeader = `
-
-###############################################################################
-#                              SERVICE INPUTS                                 #
-###############################################################################
+#
+# SERVICE INPUTS:
+#
 `
 
 // PrintSampleConfig prints the sample config
@@ -421,9 +423,9 @@ func (c *Config) addOutput(name string, table *ast.Table) error {
 
 	ro := internal_models.NewRunningOutput(name, output, outputConfig)
 	if c.Agent.MetricBufferLimit > 0 {
-		ro.PointBufferLimit = c.Agent.MetricBufferLimit
+		ro.MetricBufferLimit = c.Agent.MetricBufferLimit
 	}
-	ro.Quiet = c.Agent.Quiet
+	ro.FlushBufferWhenFull = c.Agent.FlushBufferWhenFull
 	c.Outputs = append(c.Outputs, ro)
 	return nil
 }
@@ -472,18 +474,19 @@ func (c *Config) addInput(name string, table *ast.Table) error {
 	return nil
 }
 
-// buildFilter builds a Filter (tagpass/tagdrop/pass/drop) to
+// buildFilter builds a Filter
+// (tagpass/tagdrop/namepass/namedrop/fieldpass/fielddrop) to
 // be inserted into the internal_models.OutputConfig/internal_models.InputConfig to be used for prefix
 // filtering on tags and measurements
 func buildFilter(tbl *ast.Table) internal_models.Filter {
 	f := internal_models.Filter{}
 
-	if node, ok := tbl.Fields["pass"]; ok {
+	if node, ok := tbl.Fields["namepass"]; ok {
 		if kv, ok := node.(*ast.KeyValue); ok {
 			if ary, ok := kv.Value.(*ast.Array); ok {
 				for _, elem := range ary.Value {
 					if str, ok := elem.(*ast.String); ok {
-						f.Pass = append(f.Pass, str.Value)
+						f.NamePass = append(f.NamePass, str.Value)
 						f.IsActive = true
 					}
 				}
@@ -491,13 +494,45 @@ func buildFilter(tbl *ast.Table) internal_models.Filter {
 		}
 	}
 
-	if node, ok := tbl.Fields["drop"]; ok {
+	if node, ok := tbl.Fields["namedrop"]; ok {
 		if kv, ok := node.(*ast.KeyValue); ok {
 			if ary, ok := kv.Value.(*ast.Array); ok {
 				for _, elem := range ary.Value {
 					if str, ok := elem.(*ast.String); ok {
-						f.Drop = append(f.Drop, str.Value)
+						f.NameDrop = append(f.NameDrop, str.Value)
 						f.IsActive = true
+					}
+				}
+			}
+		}
+	}
+
+	fields := []string{"pass", "fieldpass"}
+	for _, field := range fields {
+		if node, ok := tbl.Fields[field]; ok {
+			if kv, ok := node.(*ast.KeyValue); ok {
+				if ary, ok := kv.Value.(*ast.Array); ok {
+					for _, elem := range ary.Value {
+						if str, ok := elem.(*ast.String); ok {
+							f.FieldPass = append(f.FieldPass, str.Value)
+							f.IsActive = true
+						}
+					}
+				}
+			}
+		}
+	}
+
+	fields = []string{"drop", "fielddrop"}
+	for _, field := range fields {
+		if node, ok := tbl.Fields[field]; ok {
+			if kv, ok := node.(*ast.KeyValue); ok {
+				if ary, ok := kv.Value.(*ast.Array); ok {
+					for _, elem := range ary.Value {
+						if str, ok := elem.(*ast.String); ok {
+							f.FieldDrop = append(f.FieldDrop, str.Value)
+							f.IsActive = true
+						}
 					}
 				}
 			}
@@ -542,6 +577,10 @@ func buildFilter(tbl *ast.Table) internal_models.Filter {
 		}
 	}
 
+	delete(tbl.Fields, "namedrop")
+	delete(tbl.Fields, "namepass")
+	delete(tbl.Fields, "fielddrop")
+	delete(tbl.Fields, "fieldpass")
 	delete(tbl.Fields, "drop")
 	delete(tbl.Fields, "pass")
 	delete(tbl.Fields, "tagdrop")
@@ -710,6 +749,13 @@ func buildOutput(name string, tbl *ast.Table) (*internal_models.OutputConfig, er
 	oc := &internal_models.OutputConfig{
 		Name:   name,
 		Filter: buildFilter(tbl),
+	}
+	// Outputs don't support FieldDrop/FieldPass, so set to NameDrop/NamePass
+	if len(oc.Filter.FieldDrop) > 0 {
+		oc.Filter.NameDrop = oc.Filter.FieldDrop
+	}
+	if len(oc.Filter.FieldPass) > 0 {
+		oc.Filter.NamePass = oc.Filter.FieldPass
 	}
 	return oc, nil
 }